Renamed framework package eventbus into eventing
